Added user and master features
The user can see both the comments button and the master's button. However, the user can never access the features of the admin's button. This is because it is done by php. The admin's email must be masteraccount@gmail.com, in order to access the admin features, such as adding photos.